Dev - Refactored SWAL2 for easier invocation
